+++ src/QueryCat.Backend/Commands/Select/Inputs/SetKeysRowsInput.cs
@@ -1,3 +1,4 @@
+using System.Runtime.CompilerServices;
 using Microsoft.Extensions.Logging;
 using QueryCat.Backend.Commands.Select.KeyConditionValue;
 using QueryCat.Backend.Core;
@@ -8,11 +9,14 @@ using QueryCat.Backend.Utils;
 
 namespace QueryCat.Backend.Commands.Select.Inputs;
 
-internal sealed class SetKeysRowsInput : RowsInput, IRowsInputUpdate, IRowsInputDelete
+internal sealed class SetKeysRowsInput : IRowsInputUpdate, IRowsInputDelete
 {
     private sealed record ConditionJoint(
         SelectQueryCondition Condition,
-        int ColumnIndex);
+        int ColumnIndex)
+    {
+        public VariantValue? KeyValue { get; set; }
+    }
 
     private readonly int _id = IdGenerator.GetNext();
 
@@ -24,17 +28,19 @@ internal sealed class SetKeysRowsInput : RowsInput, IRowsInputUpdate, IRowsInput
     // Special mode for "WHERE id IN (x, y, z)" condition.
     private bool _hasMultipleConditions;
     private bool _hasNoMoreData;
+    private bool _needFillConditions = true;
+
     private readonly ILogger _logger = Application.LoggerFactory.CreateLogger(nameof(SetKeysRowsInput));
 
     /// <inheritdoc />
-    public override QueryContext QueryContext
+    public QueryContext QueryContext
     {
         get => _rowsInput.QueryContext;
         set => _rowsInput.QueryContext = value;
     }
 
     /// <inheritdoc />
-    public override Column[] Columns { get; protected set; }
+    public Column[] Columns { get; }
 
     public IRowsInput InnerRowsInput => _rowsInput;
 
@@ -47,13 +53,13 @@ internal sealed class SetKeysRowsInput : RowsInput, IRowsInputUpdate, IRowsInput
     }
 
     /// <inheritdoc />
-    public override Task OpenAsync(CancellationToken cancellationToken = default) => _rowsInput.OpenAsync(cancellationToken);
+    public Task OpenAsync(CancellationToken cancellationToken = default) => _rowsInput.OpenAsync(cancellationToken);
 
     /// <inheritdoc />
-    public override Task CloseAsync(CancellationToken cancellationToken = default) => _rowsInput.CloseAsync(cancellationToken);
+    public Task CloseAsync(CancellationToken cancellationToken = default) => _rowsInput.CloseAsync(cancellationToken);
 
     /// <inheritdoc />
-    public override async Task ResetAsync(CancellationToken cancellationToken = default)
+    public async Task ResetAsync(CancellationToken cancellationToken = default)
     {
         foreach (var condition in _conditions)
         {
@@ -68,17 +74,51 @@ internal sealed class SetKeysRowsInput : RowsInput, IRowsInputUpdate, IRowsInput
     }
 
     /// <inheritdoc />
-    public override ErrorCode ReadValue(int columnIndex, out VariantValue value) => _rowsInput.ReadValue(columnIndex, out value);
+    public string[] UniqueKey => _rowsInput.UniqueKey;
 
     /// <inheritdoc />
-    public override async ValueTask<bool> ReadNextAsync(CancellationToken cancellationToken = default)
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public ErrorCode ReadValue(int columnIndex, out VariantValue value) => _rowsInput.ReadValue(columnIndex, out value);
+
+    /// <inheritdoc />
+    public async ValueTask<bool> ReadNextAsync(CancellationToken cancellationToken = default)
     {
+        // Fast path.
         if (_hasNoMoreData)
         {
             return false;
         }
-        await base.ReadNextAsync(cancellationToken);
+        if (_needFillConditions)
+        {
+            FillConditions();
+            _needFillConditions = false;
+        }
+        if (_conditions.Length == 0)
+        {
+            return await _rowsInput.ReadNextAsync(cancellationToken);
+        }
 
+        // Fill conditions and check them.
+        var keysValid = false;
+        while (!keysValid)
+        {
+            var hasData = await ReadNextInternalAsync(cancellationToken);
+            if (!hasData)
+            {
+                break;
+            }
+            keysValid = CheckConditions();
+            if (keysValid)
+            {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    private async ValueTask<bool> ReadNextInternalAsync(CancellationToken cancellationToken)
+    {
         if (_conditions.Length > 0 && !_keysFilled)
         {
             await FillKeysAsync(cancellationToken);
@@ -113,10 +153,10 @@ internal sealed class SetKeysRowsInput : RowsInput, IRowsInputUpdate, IRowsInput
             {
                 hasMoreMultipleValues = await multipleValuesGenerator.MoveNextAsync(_thread, cancellationToken);
             }
-            var nullableValue = await conditionJoint.Condition.Generator.GetAsync(_thread, cancellationToken);
-            if (nullableValue.HasValue)
+            conditionJoint.KeyValue = await conditionJoint.Condition.Generator.GetAsync(_thread, cancellationToken);
+            if (conditionJoint.KeyValue.HasValue)
             {
-                _rowsInput.SetKeyColumnValue(conditionJoint.ColumnIndex, nullableValue.Value, conditionJoint.Condition.Operation);
+                _rowsInput.SetKeyColumnValue(conditionJoint.ColumnIndex, conditionJoint.KeyValue.Value, conditionJoint.Condition.Operation);
             }
             else
             {
@@ -127,8 +167,7 @@ internal sealed class SetKeysRowsInput : RowsInput, IRowsInputUpdate, IRowsInput
         return hasMoreMultipleValues;
     }
 
-    /// <inheritdoc />
-    protected override ValueTask LoadAsync(CancellationToken cancellationToken = default)
+    private void FillConditions()
     {
         // Find all related conditions.
         var list = new List<ConditionJoint>();
@@ -136,9 +175,10 @@ internal sealed class SetKeysRowsInput : RowsInput, IRowsInputUpdate, IRowsInput
         {
             foreach (var inputKeyCondition in condition.Conditions)
             {
-                list.Add(new ConditionJoint(
+                var conditionJoint = new ConditionJoint(
                     inputKeyCondition,
-                    condition.KeyColumn.ColumnIndex));
+                    condition.KeyColumn.ColumnIndex);
+                list.Add(conditionJoint);
             }
         }
         _conditions = list.ToArray();
@@ -147,8 +187,42 @@ internal sealed class SetKeysRowsInput : RowsInput, IRowsInputUpdate, IRowsInput
         _hasMultipleConditions = _conditions
             .Any(c => c.Condition.Operation == VariantValue.Operation.Equals
                         && c.Condition.Generator is IKeyConditionMultipleValuesGenerator);
+    }
 
-        return base.LoadAsync(cancellationToken);
+    /*
+     * We can check certain condition here to avoid unnecessary data reads. Example:
+     * "select * from x() inner join y() on x.id = y.id where x like '%cat%'"
+     * By current logic we will fetch all y() related records and only after that check for pattern match.
+     * With this check we verify conditions before y() fetch and reduce data load:
+     * 1. load x() rows;
+     * 2. filter x() rows; <-
+     * 3. load all related y() rows;
+     * 4. filter whole row by pattern;
+     */
+
+    private bool CheckConditions()
+    {
+        foreach (var condition in _conditions)
+        {
+            if (!condition.KeyValue.HasValue || !IsLogicalOperation(condition.Condition.Operation))
+            {
+                continue;
+            }
+            var errorCode = ReadValue(condition.ColumnIndex, out var columnValue);
+            if (errorCode != ErrorCode.OK)
+            {
+                continue;
+            }
+            var operationDelegate = VariantValue.GetOperationDelegate(
+                condition.Condition.Operation, columnValue.Type, condition.KeyValue.Value.Type);
+            var matchCondition = operationDelegate.Invoke(columnValue, condition.KeyValue.Value);
+            if (!matchCondition.AsBoolean)
+            {
+                _logger.LogWarning($"!!!!!!! {columnValue} - {condition.Condition.Operation} - {condition.KeyValue.Value}");
+                return false;
+            }
+        }
+        return true;
     }
 
     /// <inheritdoc />
@@ -172,25 +246,42 @@ internal sealed class SetKeysRowsInput : RowsInput, IRowsInputUpdate, IRowsInput
     }
 
     /// <inheritdoc />
-    public override IReadOnlyList<KeyColumn> GetKeyColumns() => _rowsInput.GetKeyColumns();
+    public IReadOnlyList<KeyColumn> GetKeyColumns() => _rowsInput.GetKeyColumns();
 
     /// <inheritdoc />
-    public override void SetKeyColumnValue(int columnIndex, VariantValue value, VariantValue.Operation operation)
+    public void SetKeyColumnValue(int columnIndex, VariantValue value, VariantValue.Operation operation)
     {
         // Do not passthru set key value calls.
     }
 
     /// <inheritdoc />
-    public override void UnsetKeyColumnValue(int columnIndex, VariantValue.Operation operation)
+    public void UnsetKeyColumnValue(int columnIndex, VariantValue.Operation operation)
     {
         // Do not passthru key value calls.
     }
 
     /// <inheritdoc />
-    public override void Explain(IndentedStringBuilder stringBuilder)
+    public void Explain(IndentedStringBuilder stringBuilder)
     {
         stringBuilder.AppendRowsInputsWithIndent($"SetKeys (id={_id})", _rowsInput);
     }
+
+    private static bool IsLogicalOperation(VariantValue.Operation operation)
+        => operation == VariantValue.Operation.Equals
+           || operation == VariantValue.Operation.NotEquals
+           || operation == VariantValue.Operation.And
+           || operation == VariantValue.Operation.Or
+           || operation == VariantValue.Operation.GreaterOrEquals
+           || operation == VariantValue.Operation.Greater
+           || operation == VariantValue.Operation.LessOrEquals
+           || operation == VariantValue.Operation.Less
+           || operation == VariantValue.Operation.Like
+           || operation == VariantValue.Operation.Between
+           || operation == VariantValue.Operation.BetweenAnd
+           || operation == VariantValue.Operation.IsNotNull
+           || operation == VariantValue.Operation.IsNull
+           || operation == VariantValue.Operation.NotLike
+           || operation == VariantValue.Operation.NotSimilar;
 
     /// <inheritdoc />
     public override string ToString() => $"Id = {_id}, RowsInput = {_rowsInput.GetType().Name}";