+++ src/QueryCat.Backend/Storage/RowsOutputQueryContext.cs
@@ -1,11 +1,12 @@
 using QueryCat.Backend.Core.Data;
+using QueryCat.Backend.Core.Execution;
 
 namespace QueryCat.Backend.Storage;
 
 /// <summary>
 /// Context for rows output.
 /// </summary>
-internal class RowsOutputQueryContext : QueryContext
+public class RowsOutputQueryContext : QueryContext
 {
     /// <summary>
     /// Constructor.
@@ -24,5 +25,18 @@ internal class RowsOutputQueryContext : QueryContext
     /// <param name="columns">Column of output.</param>
     public RowsOutputQueryContext(Column[] columns) : this(columns, NullConfigStorage.Instance)
     {
+    }
+
+    /// <summary>
+    /// Constructor.
+    /// </summary>
+    /// <param name="columns">Column of output.</param>
+    /// <param name="executionThread">Execution thread.</param>
+    public RowsOutputQueryContext(Column[] columns, IExecutionThread<ExecutionOptions> executionThread)
+        : base(new QueryContextQueryInfo(columns))
+    {
+        ConfigStorage = executionThread.ConfigStorage;
+        PrereadRowsCount = executionThread.Options.AnalyzeRowsCount;
+        SkipIfNoColumns = executionThread.Options.SkipIfNoColumns;
     }
 }